[builder] Extract common test data
